Rename+relocate fused MoE-v2 stack to nn/ddp OLMoDDP*
Adopt the olmo-ddp naming: the fused expert-parallel MoE-v2 stack is the OLMo
multi-group DDP training stack, so the canonical names/modules become:

  MoEFusedV2TransformerBlock[Config] -> OLMoDDPTransformerBlock[Config]  (nn/ddp/block.py)
  MoEFusedV2Transformer              -> OLMoDDPModel                     (nn/ddp/model.py)
  MoEFusedV2TransformerConfig        -> OLMoDDPModelConfig               (nn/transformer/config.py)
  MoEV2TransformerTrainModule[Config]-> OLMoDDPTrainModule[Config]       (ddp_train_module.py)

nn/moe/v2/{block,model}.py are now thin back-compat shims re-exporting the old
MoEFusedV2* names as aliases; nn/transformer/config.py keeps a
MoEFusedV2TransformerConfig alias. No behavior change — this is a rename +
relocation only (the model already had apply_ddp/MultiGroup DDP support).
