+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added HuggingFace interop for the MoE-v2 model (beta): the `olmo3moe` HF model (`Olmo3MoeConfig` + `modeling_olmo3moe`, a custom transformers-style implementation since transformers has no OLMoE-MoE architecture) and olmo-core ↔ HF checkpoint conversion wired into `olmo_core.nn.hf` (`convert_olmo3moe_state_{from,to}_hf` + an `Olmo3MoeConfig` builder), including the fused `w_up_gate` routed-expert split.
 - Added `Trainer.eval_checkpoints()` (and the `checkpoints_to_eval` config/trainer field): evaluate a list of checkpoint paths/globs with the configured `EvaluatorCallback`s without training. `TrainerConfig.build()` gained an `eval_only` pass-through for a uniform calling convention with the train-module config's `build()`.
-- Added `MoEV2TransformerTrainModule` (and `MoEV2TransformerTrainModuleConfig`), the train module for the fused MoE-v2 transformer. It builds the fused MoE distributed optimizer (`MoEFusedV2OptimizerConfig`), wraps model parts in `MultiGroupDistributedDataParallel`, and supports DP/EP/TP/CP/PP parallelism.
+- Added `OLMoDDPTrainModule` (and `OLMoDDPTrainModuleConfig`), the train module for the fused MoE-v2 transformer. It builds the fused MoE distributed optimizer (`MoEFusedV2OptimizerConfig`), wraps model parts in `MultiGroupDistributedDataParallel`, and supports DP/EP/TP/CP/PP parallelism.
 - Added a custom pipeline-parallel layer (`olmo_core.train.train_module.transformer.pipeline`) with `CustomPipelineStage` and the `CustomSchedule1F1BV` / `CustomScheduleInterleaved1F1B` schedules, which re-use point-to-point receive buffers across micro-batches over a NCCL-RMA transport. Enabled via `TransformerPipelineParallelConfig.use_custom_stage_implementation` and `PipelineP2PBackend`.
 - Added `MultiGroupDistributedDataParallel` (`olmo_core.nn.parallel`), a data-parallel wrapper that accumulates gradients into flat bucket views and supports per-parameter process groups (`param_process_group_fn`), overlapped bucketed all-reduce (finalized via `finalize_grad_reduce()`), and optional fp32 gradient accumulation/reduction.
 - Added `max_checkpoints` parameter to `CheckpointerCallback` (default: 3) to limit the number of permanent checkpoints retained. Oldest checkpoints are removed automatically when the limit is exceeded. Set to `None` to keep all (previous behavior).
@@ -34,12 +34,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `olmo_core.nn.fp8_weight.FP8WeightStore` (beta), a logical weight store for fp32-main / FP8-model training that caches prequantized MXFP8 expert weights and presents a tensor-like surface to the fused optimizer.
 - Added MoE token permute/reorder kernels (`olmo_core.kernels.moe_chunk_reorder`, `moe_permute_drop`, `moe_unpermute_bwd`) for routing tokens to experts (with capacity dropping) and combining the expert outputs.
 - Added symmetric-memory and NCCL-RMA point-to-point communication kernels (`olmo_core.kernels.symm_mem_vdev2d`, `olmo_symm_mem`, `nccl_rma_p2p`) for expert-parallel all-to-all and one-sided RMA transport.
-- Added a fused Mixture-of-Experts transformer block and model (`MoEFusedV2TransformerBlock` / `MoEFusedV2Transformer`, configured via `MoEFusedV2TransformerConfig` and `MoEFusedV2TransformerBlockConfig`), wired into the transformer build dispatch (`TransformerType.moe_fused_v2` / `TransformerBlockType.moe_fused_v2`) and weight initialization (`InitMethod.init_moe_v2`). This first drop runs the single-process (no expert-parallel) forward path; the expert-parallel dispatch families are imported lazily and land in follow-up changes.
+- Added a fused Mixture-of-Experts transformer block and model (`OLMoDDPTransformerBlock` / `OLMoDDPModel` in `olmo_core.nn.ddp`, configured via `OLMoDDPModelConfig` and `OLMoDDPTransformerBlockConfig`; the former `MoEFusedV2*` names remain importable as aliases), wired into the transformer build dispatch (`TransformerType.moe_fused_v2` / `TransformerBlockType.moe_fused_v2`) and weight initialization (`InitMethod.init_moe_v2`). This first drop runs the single-process (no expert-parallel) forward path; the expert-parallel dispatch families are imported lazily and land in follow-up changes.
 
 
 ### Fixed
 
-- `EvaluatorCallback` and `BatchSizeSchedulerCallback` now recognize `MoEV2TransformerTrainModule` (previously they only accepted `TransformerTrainModule`/`TransformerPipelineTrainModule`, so in-loop eval and batch-size scheduling aborted with a MoE-v2 run). `BatchSizeSchedulerCallback` also accepts the `MoEFusedV2Optimizer` for its learning-rate adjustment.
+- `EvaluatorCallback` and `BatchSizeSchedulerCallback` now recognize `OLMoDDPTrainModule` (previously they only accepted `TransformerTrainModule`/`TransformerPipelineTrainModule`, so in-loop eval and batch-size scheduling aborted with a MoE-v2 run). `BatchSizeSchedulerCallback` also accepts the `MoEFusedV2Optimizer` for its learning-rate adjustment.
 - The CPU `Test` CI job now caches `HF_HOME` across runs so the HuggingFace roundtrip tests (Qwen3-0.6B, Gemma-3-270m) don't re-download their checkpoints every run.
 - Excluded `mark_dynamic` from `torch.compile` tracing (`@torch.compiler.disable`).
 - Clearer error messages (now include the offending values) when a rank batch size isn't divisible by the sequence length, or `max_target_sequence_length` isn't a multiple of `sequence_length`.
