+++ src/test/train/train_module/transformer/ddp_train_module_test.py
@@ -1,4 +1,4 @@
-"""Tests for :class:`MoEV2TransformerTrainModule` config and construction."""
+"""Tests for :class:`OLMoDDPTrainModule` config and construction."""
 
 import torch
 
@@ -17,7 +17,7 @@ from olmo_core.nn.transformer import (
 )
 from olmo_core.optim import MoEFusedV2OptimizerConfig
 from olmo_core.testing import requires_multi_gpu, run_distributed_test
-from olmo_core.train.train_module import MoEV2TransformerTrainModuleConfig
+from olmo_core.train.train_module import OLMoDDPTrainModuleConfig
 from olmo_core.train.train_module.transformer import (
     TransformerDataParallelConfig,
     TransformerExpertParallelConfig,
@@ -26,18 +26,18 @@ from olmo_core.train.train_module.transformer import (
 
 
 def test_moe_v2_train_module_config_roundtrips():
-    config = MoEV2TransformerTrainModuleConfig(
+    config = OLMoDDPTrainModuleConfig(
         rank_microbatch_size=1024,
         max_sequence_length=512,
         optim=MoEFusedV2OptimizerConfig(lr=1e-3),
     )
-    restored = MoEV2TransformerTrainModuleConfig.from_dict(config.as_dict())
+    restored = OLMoDDPTrainModuleConfig.from_dict(config.as_dict())
     assert restored == config
     assert restored.optim.lr == 1e-3
 
 
 def test_moe_v2_train_module_config_roundtrips_with_parallelism():
-    config = MoEV2TransformerTrainModuleConfig(
+    config = OLMoDDPTrainModuleConfig(
         rank_microbatch_size=1024,
         max_sequence_length=512,
         optim=MoEFusedV2OptimizerConfig(lr=1e-3),
@@ -46,7 +46,7 @@ def test_moe_v2_train_module_config_roundtrips_with_parallelism():
         ),
         pp_config=TransformerPipelineParallelConfig(degree=2),
     )
-    restored = MoEV2TransformerTrainModuleConfig.from_dict(config.as_dict())
+    restored = OLMoDDPTrainModuleConfig.from_dict(config.as_dict())
     assert restored == config
     assert restored.dp_config is not None and restored.dp_config.reduce_grads_in_fp32 is False
     assert restored.pp_config is not None and restored.pp_config.degree == 2
@@ -83,7 +83,7 @@ def _tiny_model_config(
 
 def _run_construct_no_ep():
     model = _tiny_model_config().build(init_device="cpu")
-    config = MoEV2TransformerTrainModuleConfig(
+    config = OLMoDDPTrainModuleConfig(
         rank_microbatch_size=512,
         max_sequence_length=512,
         optim=MoEFusedV2OptimizerConfig(lr=1e-3),
@@ -112,7 +112,7 @@ def _run_construct_ep():
     # bf16 params → the fused optimizer maintains fp32 master params (its realistic config); a pure
     # fp32 model instead takes the optimizer's "expect fp32 param" branch.
     model = _tiny_model_config(dtype=DType.bfloat16).build(init_device="cuda")
-    config = MoEV2TransformerTrainModuleConfig(
+    config = OLMoDDPTrainModuleConfig(
         rank_microbatch_size=512,
         max_sequence_length=512,
         optim=MoEFusedV2OptimizerConfig(lr=1e-3),
